Update ai_connect4.kt
Updated the control flow of the game. The main menu is updated with three options Player vs AI, Player1 vs Player2 and Exit. And the post-game menu is added. After a game is finished, the player has an option to play again, return to main menu or exit. Option, play again, will return back to the game selected originally.